requirements.txt
attrs
pylint